fix use astro actions
