+++ src/actions/index.ts
@@ -1,0 +1,5 @@
+import { blogs } from "./dashboard/blogs";
+
+export const server = {
+  blogs,
+};
